+++ app/Http/Controllers/R22ReviewerEclerePenelitianMhsController.php
@@ -8,30 +8,37 @@ use App\Models\Periode;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class R22ReviewerEclerePenelitianMhsController extends Controller
 {
     public function index(Request $request, Pegawai $pegawai){
+        if (!Gate::allows('read-r022-reviewer-eclere-penelitian-mhs')) {
+            abort(403);
+        }
         $pegawais = Pegawai::all();
         $r022reviewereclerepenelitianmhs = R022ReviewerEclerePenelitianMhs::orderBy('created_at','desc')->get();
         $periode = Periode::select('nama_periode')->where('is_active','1')->first();
 
         return view('backend/rubriks/r_022_reviewer_eclere_penelitian_mhs.index',[
-           'pegawais'                              =>  $pegawais,
-           'periode'                               =>  $periode,
+           'pegawais'                           =>  $pegawais,
+           'periode'                            =>  $periode,
            'r022reviewereclerepenelitianmhs'    =>  $r022reviewereclerepenelitianmhs,
        ]);
    }
 
    public function store(Request $request){
+    if (!Gate::allows('store-r022-reviewer-eclere-penelitian-mhs')) {
+        abort(403);
+    }
        $rules = [
-           'nip'                   =>  'required|numeric',
-           'judul_protokol_penelitian'          =>  'required',
+           'nip'                        =>  'required|numeric',
+           'judul_protokol_penelitian'  =>  'required',
        ];
        $text = [
-           'nip.required'              => 'NIP harus dipilih',
-           'nip.numeric'               => 'NIP harus berupa angka',
-           'judul_protokol_penelitian.required'     => 'Jumlah mhs harus diisi',
+           'nip.required'                       => 'NIP harus dipilih',
+           'nip.numeric'                        => 'NIP harus berupa angka',
+           'judul_protokol_penelitian.required' => 'Judul Protokol Penelitian harus diisi',
        ];
 
        $validasi = Validator::make($request->all(), $rules, $text);
@@ -59,18 +66,24 @@ class R22ReviewerEclerePenelitianMhsController extends Controller
        }
    }
    public function edit(R022ReviewerEclerePenelitianMhs $r22revieweclerepenelitimhs){
+    if (!Gate::allows('edit-r022-reviewer-eclere-penelitian-mhs')) {
+        abort(403);
+    }
        return $r22revieweclerepenelitimhs;
    }
 
    public function update(Request $request, R022ReviewerEclerePenelitianMhs $r22revieweclerepenelitimhs){
+    if (!Gate::allows('update-r022-reviewer-eclere-penelitian-mhs')) {
+        abort(403);
+    }
        $rules = [
-           'nip'                   =>  'required|numeric',
-           'judul_protokol_penelitian'          =>  'required',
+           'nip'                        =>  'required|numeric',
+           'judul_protokol_penelitian'  =>  'required',
        ];
        $text = [
-           'nip.required'          => 'NIP harus dipilih',
-           'nip.numeric'           => 'NIP harus berupa angka',
-           'judul_protokol_penelitian.required' => 'Jumlah mhs harus diisi',
+           'nip.required'                       => 'NIP harus dipilih',
+           'nip.numeric'                        => 'NIP harus berupa angka',
+           'judul_protokol_penelitian.required' => 'Judul Protokol Penelitian harus diisi',
        ];
 
        $validasi = Validator::make($request->all(), $rules, $text);
@@ -98,6 +111,9 @@ class R22ReviewerEclerePenelitianMhsController extends Controller
        }
    }
    public function delete(R022ReviewerEclerePenelitianMhs $r22revieweclerepenelitimhs){
+    if (!Gate::allows('delete-r022-reviewer-eclere-penelitian-mhs')) {
+        abort(403);
+    }
        $delete = $r22revieweclerepenelitimhs->delete();
        if ($delete) {
            $notification = array(
