+++ app/Http/Controllers/R24TimAkredProdiDanDirektoratController.php
@@ -8,10 +8,14 @@ use App\Models\Periode;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class R24TimAkredProdiDanDirektoratController extends Controller
 {
     public function index(Request $request, Pegawai $pegawai){
+        if (!Gate::allows('read-r024-tim-akred-prodi-dan-direktorat')) {
+            abort(403);
+        }
         $pegawais = Pegawai::all();
         $r024timakredprodirektorats = R024TimAkredProdiDanDirektorat::orderBy('created_at','desc')->get();
         $periode = Periode::select('nama_periode')->where('is_active','1')->first();
@@ -24,6 +28,9 @@ class R24TimAkredProdiDanDirektoratController extends Controller
    }
 
    public function store(Request $request){
+    if (!Gate::allows('store-r024-tim-akred-prodi-dan-direktorat')) {
+        abort(403);
+    }
        $rules = [
            'nip'                   =>  'required|numeric',
            'judul_kegiatan'          =>  'required',
@@ -59,10 +66,16 @@ class R24TimAkredProdiDanDirektoratController extends Controller
        }
    }
    public function edit(R024TimAkredProdiDanDirektorat $r24timakredprodirektorat){
+    if (!Gate::allows('edit-r024-tim-akred-prodi-dan-direktorat')) {
+        abort(403);
+    }
        return $r24timakredprodirektorat;
    }
 
    public function update(Request $request, R024TimAkredProdiDanDirektorat $r24timakredprodirektorat){
+    if (!Gate::allows('update-r024-tim-akred-prodi-dan-direktorat')) {
+        abort(403);
+    }
        $rules = [
            'nip'                     =>  'required|numeric',
            'judul_kegiatan'          =>  'required',
@@ -98,6 +111,9 @@ class R24TimAkredProdiDanDirektoratController extends Controller
        }
    }
    public function delete(R024TimAkredProdiDanDirektorat $r24timakredprodirektorat){
+    if (!Gate::allows('delete-r024-tim-akred-prodi-dan-direktorat')) {
+        abort(403);
+    }
        $delete = $r24timakredprodirektorat->delete();
        if ($delete) {
            $notification = array(
